+++ sentry/src/main/java/io/sentry/IScope.java
@@ -2,6 +2,7 @@ package io.sentry;
 
 import io.sentry.protocol.Contexts;
 import io.sentry.protocol.Request;
+import io.sentry.protocol.SentryId;
 import io.sentry.protocol.User;
 import java.util.Collection;
 import java.util.List;
@@ -83,6 +84,23 @@ public interface IScope {
    */
   @ApiStatus.Internal
   void setScreen(final @Nullable String screen);
+
+  /**
+   * Returns the Scope's current replay_id, previously set by {@link IScope#setReplayId(SentryId)}
+   *
+   * @return the id of the current session replay
+   */
+  @ApiStatus.Internal
+  @NotNull
+  SentryId getReplayId();
+
+  /**
+   * Sets the Scope's current replay_id
+   *
+   * @param replayId the id of the current session replay
+   */
+  @ApiStatus.Internal
+  void setReplayId(final @NotNull SentryId replayId);
 
   /**
    * Returns the Scope's request