+++ sentry/src/main/java/io/sentry/SentryItemType.java
@@ -18,6 +18,7 @@ public enum SentryItemType implements JsonSerializable {
   ClientReport("client_report"),
   ReplayEvent("replay_event"),
   ReplayRecording("replay_recording"),
+  ReplayVideo("replay_video"),
   CheckIn("check_in"),
   Statsd("statsd"),
   Unknown("__unknown__"); // DataCategory.Unknown
@@ -65,7 +66,7 @@ public enum SentryItemType implements JsonSerializable {
 
     @Override
     public @NotNull SentryItemType deserialize(
-        @NotNull JsonObjectReader reader, @NotNull ILogger logger) throws Exception {
+        @NotNull ObjectReader reader, @NotNull ILogger logger) throws Exception {
       return SentryItemType.valueOfLabel(reader.nextString().toLowerCase(Locale.ROOT));
     }
   }